Решение заданиia 1

+++ src/exercise1.c
@@ -2,13 +2,13 @@
 #include<stdlib.h>
 
 int main(int argc,char *argv[]){
-    int a[10];
+    double a[10];
     for(int i = 0; i<10; i++){
-        a[i] = atoi(argv[i+1]);
+        a[i] = atof(argv[i+1]);
     }
-    float c = 0;
+    double c = 0;
     for(int i = 0; i<10; i++){
         c+=a[i];
     }
-    printf("%.2f", c/10);
+    printf("%.2f", c/10.0);
 }
